requirements.txt
flask>=3.0.0
mysql-connector-python>=8.0.0
fpdf2>=2.7.0
gunicorn>=21.0.0
waitress>=3.0.0
python-dotenv>=1.0.0
